+++ envswitch/tests/test_gui_config.py
@@ -3,6 +3,7 @@ import tempfile
 
 import pytest
 
+from envswitch.gui import EnvSwitcherApp, EnvSwitcherAppHeadless
 from envswitch.env_config import GlobalEnvsConfig
 
 THIS_DIR = os.path.dirname(os.path.abspath(__file__))
@@ -37,3 +38,16 @@ def test_load_gui_config(config_file_name: str):
         os.close(fd)
         os.remove(fpath)
 
+
+@pytest.mark.parametrize("config_file_name", os.listdir(os.path.join(THIS_DIR, 'data')))
+def test_list_available_envs(config_file_name: str):
+    conf_file_path = os.path.join(THIS_DIR, 'data', config_file_name)
+
+    # first load the configuration
+    with open(conf_file_path, 'r') as f:
+        conf = GlobalEnvsConfig.from_yaml(f)
+        assert type(conf) == GlobalEnvsConfig
+
+    # assert that the list of available environments is the same when we use the app to open it
+    app = EnvSwitcherAppHeadless(config_file_path=conf_file_path)
+    assert conf.get_available_envs() == app.get_current_config().get_available_envs()